my-production-agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
redis==5.0.8
pydantic-settings==2.10.1
openai>=2.26.0,<3.0.0

